Create Divide the apples
n schoolchildren divide k apples evenly, the residue remains in the basket. How many apples remains in the basket?

Input
Two integers n and k not greater than 1500 --- rarely happens in school more pupils, and where to find such a basket?

Output
Print the number of apples in the basket.